refactor: collapse ack module into hl7v2
