+++ crates/hl7v2-ack/Cargo.toml
@@ -13,12 +13,12 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-hl7v2-core = { version = "1.2.0", path = "../hl7v2-core" }
-chrono = { version = "0.4.44", features = ["serde", "std", "clock"], default-features = false }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = [
+    "ack",
+] }
 
 [dev-dependencies]
 hl7v2-test-utils = { path = "../hl7v2-test-utils" }
-hl7v2-writer = { version = "1.2.0", path = "../hl7v2-writer" }
 proptest = "1.6"
 cucumber = "0.22.1"
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
@@ -26,4 +26,3 @@ tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 [[test]]
 name = "bdd_tests"
 harness = false
-
