+++ crates/hl7v2/Cargo.toml
@@ -28,7 +28,7 @@ profile = [
     "dep:chrono",
     "dep:regex",
 ]
-ack = ["dep:hl7v2-ack"]
+ack = ["dep:chrono"]
 normalize = ["dep:hl7v2-normalize"]
 batch = []
 stream = ["dep:hl7v2-stream"]
@@ -68,7 +68,6 @@ uuid = { workspace = true, features = ["v4"], optional = true }
 
 hl7v2-json = { version = "1.2.0", path = "../hl7v2-json", optional = true }
 hl7v2-normalize = { version = "1.2.0", path = "../hl7v2-normalize", optional = true }
-hl7v2-ack = { version = "1.2.0", path = "../hl7v2-ack", optional = true }
 hl7v2-prof = { version = "1.2.0", path = "../hl7v2-prof", optional = true }
 hl7v2-validation = { version = "1.2.0", path = "../hl7v2-validation", optional = true }
 hl7v2-stream = { version = "1.2.0", path = "../hl7v2-stream", optional = true }
